+++ zarrs/src/array/codec/array_to_bytes/sharding/sharding_partial_decoder_sync.rs
@@ -635,8 +635,8 @@ pub(crate) fn partial_decode_variable_indexer(
 
         let (element_bytes, element_offsets) = inner_partial_decoder
             .partial_decode(&[indices_in_inner_chunk], options)?
-            .into_variable()
-            .expect("fixed data");
+            .into_variable()?
+            .into_parts();
         debug_assert_eq!(element_offsets.len(), 2);
         bytes.extend_from_slice(&element_bytes);
         offsets.push(bytes.len());